requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
SQLAlchemy==1.4.25
pyodbc==4.0.32
faiss-cpu==1.7.1
sentence-transformers==2.2.2
Jinja2==3.0.1
python-dotenv==0.19.1
